+++ src/feature/schedule/ui/ReservationBottomSheet.tsx
@@ -6,9 +6,8 @@ import { useQueryClient } from '@tanstack/react-query';
 import dayjs from 'dayjs';
 import { useState } from 'react';
 
-import { IconCheck } from '@/shared/assets';
+import { IconCheck, IconError } from '@/shared/assets';
 import CancelCalendarIcon from '@/shared/assets/images/icon_cancel_calendar.svg';
-import ErrorIcon from '@/shared/assets/images/icon_error.svg';
 import ReservationCalendarIcon from '@/shared/assets/images/icon_reservation_calendar.svg';
 import { Typography } from '@/shared/mixin';
 import {
@@ -79,7 +78,7 @@ export const ReservationBottomSheet = ({ data, date }: Props) => {
           className: 'h-12',
           description: (
             <div className='flex items-center justify-center'>
-              <ErrorIcon fill={'var(--primary-500)'} />
+              <IconError fill={'var(--primary-500)'} />
               <p className='typography-heading-5 ml-6 text-[#fff]'>
                 {error.response?.data.message}
               </p>
