+++ src/page/home/ui/TrainerHomePage.tsx
@@ -1,10 +1,12 @@
 'use client';
 
 import Image from 'next/image';
+import Link from 'next/link';
 import { useRouter } from 'next/navigation';
 
 import { useAuthAction } from '@/entity/auth';
-import { IconAlarm } from '@/shared/assets';
+import { IconAlarmWhite, IconPlus } from '@/shared/assets';
+import { Typography } from '@/shared/mixin';
 import {
   Button,
   Card,
@@ -13,35 +15,8 @@ import {
   ScrollArea,
   ScrollBar,
 } from '@/shared/ui';
-import {
-  Select,
-  SelectContent,
-  SelectGroup,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from '@/shared/ui/select'; //index에 추가하고 삭제
 import { cn } from '@/shared/utils';
 import { Layout } from '@/widget';
-
-const GYM_LIST = [
-  {
-    name: '건강해짐 홍대점',
-    id: 1,
-  },
-  {
-    name: '건강해짐 떡잎마을점',
-    id: 2,
-  },
-  {
-    name: '건강해짐 당산점',
-    id: 3,
-  },
-  {
-    name: '건강해짐 울산점',
-    id: 4,
-  },
-];
 
 // TODO) 스케줄 관리 entity
 const TODAY_CLASSES = [
@@ -78,24 +53,13 @@ const MEMBER_COUNT = 20;
 export const TrainerHomePage = () => {
   const { deleteUserInfo } = useAuthAction();
   const router = useRouter();
+  const GYM_NAME = '건강해짐 홍대점';
 
   return (
-    <Layout type='trainer'>
-      <Layout.Header className='bg-white'>
-        <Select defaultValue={'1'}>
-          <SelectTrigger className='flex w-fit items-center gap-x-[4px]'>
-            <SelectValue placeholder='헬스장 선택' />
-          </SelectTrigger>
-          <SelectContent>
-            <SelectGroup>
-              {GYM_LIST.map((item) => (
-                <SelectItem key={item.id} value={String(item.id)}>
-                  {item.name}
-                </SelectItem>
-              ))}
-            </SelectGroup>
-          </SelectContent>
-        </Select>
+    <Layout type='trainer' className='relative'>
+      <div className='absolute left-0 top-0 z-0 h-[150px] w-full bg-primary' />
+      <Layout.Header className='z-10'>
+        <p className={cn(Typography.TITLE_2, 'text-white')}>{GYM_NAME}</p>
         {/* 임시로 로그아웃 기능 적용 */}
         <Button
           variant='ghost'
@@ -103,15 +67,18 @@ export const TrainerHomePage = () => {
             deleteUserInfo();
             router.push('/');
           }}>
-          <IconAlarm />
+          <IconAlarmWhite />
         </Button>
       </Layout.Header>
       <Layout.Contents>
-        <div className=' flex w-full flex-col gap-y-5 bg-white py-6'>
-          <div className='typography-heading-2 flex gap-x-2 px-5'>
-            <p>오늘의 수업</p>
-            <span className='text-primary-500'>{TODAY_CLASSES.length}</span>
-          </div>
+        <div className=' flex w-full flex-col gap-y-5 py-6'>
+          <h3
+            className={cn(
+              Typography.HEADING_4_BOLD,
+              'z-10 flex gap-x-2 px-7 text-white'
+            )}>
+            오늘의 수업
+          </h3>
           <ScrollArea className='w-full whitespace-nowrap'>
             <div className='flex w-max space-x-2.5 px-5'>
               {TODAY_CLASSES.map((item) => {
@@ -119,7 +86,7 @@ export const TrainerHomePage = () => {
                   <div
                     key={item.userId}
                     className={cn(
-                      'h-50 w-[100px] rounded-[12px] border border-gray-300 px-[24px] py-[26px] shadow-point'
+                      'h-50 w-[100px] rounded-[12px] border border-gray-300 bg-white px-[24px] py-[26px] shadow-point'
                     )}>
                     <p className={'typography-body-1 text-gray-400'}>{item.date}</p>
                     <p className={'typography-heading-4 font-bold text-gray-800'}>
@@ -133,58 +100,46 @@ export const TrainerHomePage = () => {
           </ScrollArea>
         </div>
         <div className='mt-[19px] flex flex-col items-center px-[20px]'>
-          <div className='grid grid-cols-2 gap-[10px]'>
-            <Card
-              className='h-[140px] w-[155px] cursor-pointer'
-              onClick={() => router.push('/trainer/manage')}>
-              <CardHeader>
-                <p className='flex gap-x-[4px]'>
-                  회원<span className='text-primary-500'>{MEMBER_COUNT}</span>
-                </p>
-              </CardHeader>
-              <CardContent>
-                <p>{`간편한 회원 관리와\n운동 일지 공유`}</p>
-                <div className='absolute bottom-[13px] right-[8px] h-[60px] w-[60px]'>
-                  <Image
-                    src='/images/icon_profile_coin_shadow.png'
-                    fill
-                    sizes='auto'
-                    priority
-                    alt='Card image of account'
-                  />
-                </div>
-              </CardContent>
-            </Card>
-            <Card className='h-[140px] w-[155px]'>
-              <CardHeader>수업 일지</CardHeader>
-              <CardContent>
-                <p>{`PT 일정을\n한눈에 조회!`}</p>
-                <div className='absolute bottom-[13px] right-[8px] h-[60px] w-[60px]'>
-                  <Image
-                    src='/images/icon_calendar_shadow.png'
-                    fill
-                    sizes='auto'
-                    alt='Calendar image for classes'
-                    priority
-                  />
-                </div>
-              </CardContent>
-            </Card>
-            <Card className='h-[140px] w-[155px]'>
-              <CardHeader>피드백 작성</CardHeader>
-              <CardContent>
-                <p>{`수업 내역 관리와\n피드백 작성`}</p>
-                <div className='absolute bottom-[13px] right-[8px] h-[60px] w-[60px]'>
-                  <Image
-                    src='/images/icon_calendar_shadow.png'
-                    fill
-                    sizes='auto'
-                    alt='Calendar image for classes'
-                    priority
-                  />
-                </div>
-              </CardContent>
-            </Card>
+          <div className='grid w-full grid-cols-2 gap-2'>
+            <Link href={'/trainer/manage'}>
+              <Card className='h-[140px] w-full'>
+                <CardHeader className='relative'>
+                  회원<span className='ml-1 text-primary-500'>{MEMBER_COUNT}</span>
+                  <div className='absolute right-0 top-0'>
+                    <IconPlus width={20} fill={'#CBCFD3'} />
+                  </div>
+                </CardHeader>
+                <CardContent className='relative h-full'>
+                  <p>{`간편한 회원 관리와\n운동 일지 공유`}</p>
+                  <div className='absolute bottom-0 right-0 h-[60px] w-[60px]'>
+                    <Image
+                      src='/images/icon_profile_coin_shadow.png'
+                      fill
+                      sizes='auto'
+                      priority
+                      alt='Card image of account'
+                    />
+                  </div>
+                </CardContent>
+              </Card>
+            </Link>
+            <Link href={'#'}>
+              <Card className='h-[140px] w-full'>
+                <CardHeader>피드백 작성</CardHeader>
+                <CardContent className='relative h-full'>
+                  <p>{`수업 내역 관리와\n피드백 작성`}</p>
+                  <div className='absolute bottom-0 right-0 h-[60px] w-[60px]'>
+                    <Image
+                      src='/images/icon_calendar_shadow.png'
+                      fill
+                      sizes='auto'
+                      alt='Calendar image for classes'
+                      priority
+                    />
+                  </div>
+                </CardContent>
+              </Card>
+            </Link>
           </div>
         </div>
       </Layout.Contents>
