+++ src/page/manage/ui/StudentLogPage.tsx
@@ -5,7 +5,6 @@ import 'dayjs/locale/ko';
 import dayjs from 'dayjs';
 dayjs.locale('ko');
 
-import Image from 'next/image';
 import Link from 'next/link';
 import { usePathname } from 'next/navigation';
 import { useState } from 'react';
@@ -14,14 +13,9 @@ import { useStudentLogListQuery } from '@/feature/log';
 import { IconBack, IconCalendarX, IconChat, IconPlus } from '@/shared/assets';
 import { Typography } from '@/shared/mixin';
 import { Card, CardContent, CardFooter, CardHeader } from '@/shared/ui';
-import {
-  Carousel,
-  CarouselContent,
-  CarouselItem,
-  CarouselNav,
-} from '@/shared/ui/carousel';
 import { cn } from '@/shared/utils';
 import { Layout, MonthPicker } from '@/widget';
+import { ImageSlide } from '@/widget/image-slide';
 
 interface Props {
   memberId: number;
@@ -51,58 +45,39 @@ const StudentLogPage = ({ memberId }: Props) => {
           <IconPlus fill='black' width={20} height={20} />
         </Link>
       </Layout.Header>
-      <Layout.Contents className='overflow-y-hidden py-[20px]'>
-        <div className='px-[20px]'>
+      <Layout.Contents className='overflow-y-hidden py-7'>
+        <div className='px-7'>
           <MonthPicker
             date={selectedMonth}
             onChangeDate={(newDate) => setSelectedMonth(newDate)}
           />
         </div>
         {contents && contents.length > 0 && (
-          <div className='mt-[4px] flex h-full flex-1 flex-grow flex-col overflow-y-auto px-[20px] pb-[20px]'>
-            <div className='flex w-full flex-col gap-y-[16px] pb-[16px]'>
+          <div className='mt-1 flex h-full flex-1 flex-grow flex-col overflow-y-auto px-7 pb-7'>
+            <div className='flex w-full flex-col gap-y-6 pb-6'>
               {contents.length > 0 &&
                 contents.map((log) => {
                   const date = dayjs(log.createdAt);
                   const formattedDate = date.format('M월 D일 (ddd)');
                   return (
                     <Link key={log.id} href={`${pathname}/${log.id}`}>
-                      <Card className='w-full px-0'>
-                        <CardHeader
-                          className={cn(Typography.TITLE_3, 'px-[16px] text-gray-600')}>
+                      <Card className='w-full gap-0 px-6 py-7'>
+                        <CardHeader className={cn(Typography.TITLE_3)}>
                           {formattedDate}
                           {` `}
                           {log.lessonTime}
                         </CardHeader>
-                        <CardContent className='px-[16px]'>
-                          <Carousel className='mx-auto w-full' opts={{}}>
-                            <CarouselContent className='mx-0 p-0'>
-                              {log.files.map((file, index) => (
-                                <CarouselItem key={index}>
-                                  <Card className='w-full'>
-                                    <CardContent>
-                                      <Image
-                                        src={file.fileUrl}
-                                        alt={'React Rendezvous'}
-                                        width={300}
-                                        height={200}
-                                        className={cn(
-                                          'rounded-lg object-contain transition-all',
-                                          'aspect-[3/2]' // 'aspect-square' // 1:1
-                                        )}
-                                      />
-                                    </CardContent>
-                                  </Card>
-                                </CarouselItem>
-                              ))}
-                            </CarouselContent>
-                            <CarouselNav />
-                          </Carousel>
-                          <p className={cn(Typography.BODY_3, 'py-[8px] text-black')}>
+                        <CardContent className='mt-4'>
+                          <ImageSlide images={log.files} />
+                          <p
+                            className={cn(
+                              Typography.BODY_3,
+                              'mt-5 line-clamp-2 h-full overflow-ellipsis text-black'
+                            )}>
                             {log.content}
                           </p>
                         </CardContent>
-                        <CardFooter className='flex items-center space-x-[6px] px-[16px]'>
+                        <CardFooter className='mt-6 flex items-center gap-2'>
                           <IconChat />
                           <p className={cn(Typography.BODY_4_MEDIUM, 'text-gray-500')}>
                             댓글 <span>{log.commentTotalCount}</span>
@@ -116,7 +91,7 @@ const StudentLogPage = ({ memberId }: Props) => {
           </div>
         )}
         {contents && contents.length === 0 && (
-          <div className='flex h-full flex-col items-center justify-center space-y-[10px]'>
+          <div className='flex h-full flex-col items-center justify-center space-y-4'>
             <IconCalendarX />
             <p className={cn(Typography.HEADING_4_SEMIBOLD, 'text-gray-400')}>
               수업일지 내역이 없습니다.
