+++ src/shared/assets/index.ts
@@ -1,4 +1,5 @@
 export { default as IconAlarm } from './images/alarm.svg';
+export { default as IconAlarmWhite } from './images/alarm_white.svg';
 export { default as IconArrowDown } from './images/arrow_down.svg';
 export { default as IconArrowFilledDown } from './images/arrow_filled_down.svg';
 export { default as IconArrowFilledUp } from './images/arrow_filled_up.svg';
@@ -19,6 +20,8 @@ export { default as IconClose } from './images/close.svg';
 export { default as IconCommunityFilled } from './images/community_filled.svg';
 export { default as IconCommunityOutlined } from './images/community_outlined.svg';
 export { default as IconDocumentProfile } from './images/document_profile.svg';
+export { default as IconDotsVertical } from './images/dots_vertical.svg';
+export { default as IconEdit } from './images/edit.svg';
 export { default as IconError } from './images/error.svg';
 export { default as IconGear } from './images/gear.svg';
 export { default as IconHomeFilled } from './images/home_filled.svg';
@@ -32,3 +35,4 @@ export { default as IconPlus } from './images/plus.svg';
 export { default as IconPoint } from './images/point.svg';
 export { default as IconProfileFilled } from './images/profile_filled.svg';
 export { default as IconProfileOutlined } from './images/profile_outlined.svg';
+export { default as IconTrash } from './images/trash.svg';
